+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 ## Version 1.0 (14/04/2020)
 
 Release of a first stable version
+
+## Version 2.0 (10/03/2021)
+
+Release of a new version supported by Maltego Technologies GmbH.  
+The entities and transforms are available in the official Transform Hub.  
+The python package maltego-stix2 is released to help transform developpers to use these entities in their own projects.